refactor: 移除 rust_decimal 依赖，统一使用 f64
- Position.unrealized_pnl: Decimal → f64
- Balance.available/frozen: Decimal → f64
- ExchangeError.InsufficientBalance: Decimal → f64
- FundingArbStrategyConfig: Decimal → f64
- 更新所有 codec 解析逻辑

+++ src/domain/error.rs
@@ -1,5 +1,4 @@
 use crate::domain::model::Exchange;
-use rust_decimal::Decimal;
 use std::time::Duration;
 use thiserror::Error;
 
@@ -19,7 +18,7 @@ pub enum ExchangeError {
     OrderRejected(Exchange, String),
 
     #[error("Insufficient balance on {0}: need {1}, have {2}")]
-    InsufficientBalance(Exchange, Decimal, Decimal),
+    InsufficientBalance(Exchange, f64, f64),
 
     #[error("Symbol not found on {0}: {1}")]
     SymbolNotFound(Exchange, String),
